Renaming vproc files
